remove old rpc
