--- cmd/client/keyfile/shard1-0x0a57a2714e193b7ac50475ce625f2dcfb483d741
@@ -1,10 +1,0 @@
-{
-	"version": 1,
-	"address": "0x0a57a2714e193b7ac50475ce625f2dcfb483d741",
-	"crypto": {
-		"ciphertext": "24e1147383ff30193e14a19fd34234c3309e78a863c55b5910757592d72bd066",
-		"iv": "e5c346d3c3305de6f7a21079ecf30bd1",
-		"salt": "835e6eb92bc72a6d865d52d695e89b3d50ea23673292b3851813f6fffbc591e7",
-		"mac": "0x038bea1903896694c96c3849bd0f70285e204a51cb9fcde3f8569168ddf5f8f7"
-	}
-}
