+++ cmd/util/client.go
@@ -3,81 +3,20 @@
 *  @copyright defined in go-seele/LICENSE
  */
 
-package main
+package util
 
 import (
 	"crypto/ecdsa"
 	"fmt"
 	"math/big"
 
+	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/seeleteam/go-seele/common"
-	"github.com/seeleteam/go-seele/common/hexutil"
 	"github.com/seeleteam/go-seele/core/types"
 	"github.com/seeleteam/go-seele/crypto"
 	"github.com/seeleteam/go-seele/rpc2"
+	"github.com/seeleteam/go-seele/seele"
 )
-
-const (
-	// DefaultNonce is the default value of nonce,when you are not set the nonce flag in client sendtx command by --nonce .
-	DefaultNonce uint64 = 0
-)
-
-func checkParameter(publicKey *ecdsa.PublicKey, client *rpc.Client) (*types.TransactionData, error) {
-	info := &types.TransactionData{}
-	var err error
-	if len(toValue) > 0 {
-		toAddr, err := common.HexToAddress(toValue)
-		if err != nil {
-			return info, fmt.Errorf("invalid receiver address: %s", err)
-		}
-		info.To = toAddr
-	}
-
-	amount, ok := big.NewInt(0).SetString(amountValue, 10)
-	if !ok {
-		return info, fmt.Errorf("invalid amount value")
-	}
-	info.Amount = amount
-
-	fee, ok := big.NewInt(0).SetString(feeValue, 10)
-	if !ok {
-		return info, fmt.Errorf("invalid fee value")
-	}
-	info.Fee = fee
-
-	fromAddr := crypto.GetAddress(publicKey)
-	info.From = *fromAddr
-
-	if client != nil {
-		nonce, err := GetAccountNonce(client, *fromAddr)
-		if err != nil {
-			return info, fmt.Errorf("failed to get the sender account nonce: %s", err)
-		}
-
-		if nonceValue == nonce || nonceValue == DefaultNonce {
-			info.AccountNonce = nonce
-		} else {
-			if nonceValue < nonce {
-				return info, fmt.Errorf("your nonce is: %d,current nonce is: %d,you must set your nonce greater than or equal to current nonce", nonceValue, nonce)
-			}
-			info.AccountNonce = nonceValue
-		}
-
-		fmt.Printf("account %s current nonce: %d, sending nonce: %d\n", fromAddr.ToHex(), nonce, info.AccountNonce)
-	} else {
-		info.AccountNonce = nonceValue
-	}
-
-	payload := []byte(nil)
-	if len(paloadValue) > 0 {
-		if payload, err = hexutil.HexToBytes(paloadValue); err != nil {
-			return info, fmt.Errorf("invalid payload, %s", err)
-		}
-	}
-	info.Payload = payload
-
-	return info, nil
-}
 
 // GetAccountNonce get account nonce by account
 func GetAccountNonce(client *rpc.Client, account common.Address) (uint64, error) {
@@ -86,7 +25,22 @@ func GetAccountNonce(client *rpc.Client, account common.Address) (uint64, error)
 	return nonce, err
 }
 
-func generateTx(from *ecdsa.PrivateKey, to common.Address, amount *big.Int, fee *big.Int, nonce uint64, payload []byte) (*types.Transaction, error) {
+// GetAccountNonce get account nonce by account
+func GetBalance(client *rpc.Client, account common.Address) (*big.Int, error) {
+	var result hexutil.Big
+	err := client.Call(&result, "seele_getBalance", account)
+
+	return (*big.Int)(&result), err
+}
+
+func GetInfo(client *rpc.Client) (seele.MinerInfo, error) {
+	var info seele.MinerInfo
+	err := client.Call(&info, "seele_getInfo")
+
+	return info, err
+}
+
+func GenerateTx(from *ecdsa.PrivateKey, to common.Address, amount *big.Int, fee *big.Int, nonce uint64, payload []byte) (*types.Transaction, error) {
 	fromAddr := crypto.GetAddress(&from.PublicKey)
 
 	var tx *types.Transaction
@@ -139,4 +93,17 @@ func Call(client *rpc.Client, from *ecdsa.PrivateKey, to *common.Address, amount
 	}
 
 	return &result, nil
+}
+
+func GetTransactionByHash(client *rpc.Client, hash string) (map[string]interface{}, error) {
+	var result map[string]interface{}
+	err := client.Call(&result, "txpool_getTransactionByHash", hash)
+	return result, err
+}
+
+func SendTx(client *rpc.Client, tx *types.Transaction) (bool, error) {
+	var result bool
+	err := client.Call(&result, "seele_addTx", tx)
+
+	return result, err
 }
